# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: GNNcab - Multi-city ride dispatch and driver tracking platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [audit]
      summary: Get admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [users]
      summary: Get all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, driver, customer]
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [users]
      summary: Get user by ID (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateAdminUser
      tags: [users]
      summary: Update user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [driver, customer]
      required:
        - email
        - password
        - firstName
        - lastName
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, driver, customer]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
        - updatedAt
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
    UsersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
      required:
        - data
        - total
        - page
        - limit
        - totalPages
...
    AdminStats:
      type: object
      properties:
        totalDrivers:
          type: number
        activeDrivers:
          type: number
        totalCustomers:
          type: number
        totalTrips:
          type: number
        activeTrips:
          type: number
        completedTrips:
          type: number
        cancelledTrips:
          type: number
        totalRevenue:
          type: number
        totalCities:
          type: number
        totalZones:
          type: number
      required:
        - totalDrivers
        - activeDrivers
        - totalCustomers
        - totalTrips
        - activeTrips
        - completedTrips
        - cancelledTrips
        - totalRevenue
        - totalCities
        - totalZones
...